requirements.txt
# Python 3.11+ required

# ==================== CORE FRAMEWORK ====================
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-core>=2.14.0
pydantic-settings>=2.0.0

# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-multipart>=0.0.9

# ==================== LLM & AI ====================
openai>=1.10.0
anthropic>=0.15.0
mistralai>=0.1.0

# ==================== VECTOR DB & EMBEDDINGS ====================
sentence-transformers>=2.3.0
chromadb>=0.4.22

# LangChain ecosystem (Aligned for v0.1/v0.2/v0.3 compatibility)
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
langchain-chroma>=0.1.0
langchain-huggingface>=0.0.1
langchain-community>=0.0.10
langchain>=0.1.0

# ==================== DOCUMENT PROCESSING ====================
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# ==================== TEXT PROCESSING ====================
tiktoken>=0.5.0
nltk>=3.8.1

# ==================== FRONTEND ====================
streamlit>=1.30.0
plotly>=5.18.0

# ==================== UTILITIES ====================
requests>=2.31.0
loguru>=0.7.2
redis>=5.0.0
prometheus-client>=0.19.0

# ==================== COMMON DEPENDENCIES ====================
# Allow NumPy 2.x for Python 3.13 support
numpy>=1.26.0
pandas>=2.2.0,<3.0
SQLAlchemy>=2.0.0,<3.0
aiohttp>=3.9.0,<4.0

